Split Group details page into Group Members, Group Datasources and few small components
