+++ client/app/services/navigateTo.js
@@ -1,9 +1,12 @@
 import { isString } from 'lodash';
 import { $location, $rootScope } from '@/services/ng';
 
-export default function navigateTo(url) {
+export default function navigateTo(url, replace = false) {
   if (isString(url)) {
     $location.url(url);
+    if (replace) {
+      $location.replace();
+    }
     $rootScope.$applyAsync();
   }
 }
